creating Modal component and moving the modal code from Movie Component to Modal component
