+++ src/App.js
@@ -1,6 +1,5 @@
 import './assets/css/App.css';
 
-import Movies from './components/Movies';
 import Router from './Router';
 
 function App() {
